"Refactored code: updated entity and repository names, added/removed methods, and modified service implementations."

+++ src/main/java/com/example/StockOvin/Repository/OrdersRepository.java
@@ -3,5 +3,5 @@ package com.example.StockOvin.Repository;
 import com.example.StockOvin.Entities.OrdersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdersRepository extends JpaRepository<OrdersEntity, Long> {
+public interface OrdersRepository extends JpaRepository<OrdersEntity, Integer> {
 }
